feat: move get user location from app to context file
