+++ src/index.scss
@@ -1,4 +1,5 @@
 @import 'styles/variables';
+
 *,
 *::before,
 *::after {
