+++ src/utils/location.ts
@@ -1,0 +1,4 @@
+export const getPosition = (options?: PositionOptions): Promise<GeolocationPosition> =>
+    new Promise((resolve, reject) =>
+        navigator.geolocation.getCurrentPosition(resolve, reject, options)
+    );
